shunting yard is now norm conform and as far as i tested it worked
